+++ tests/testthat/test-scenarioBuilder.R
@@ -691,4 +691,123 @@ test_that("Add new 'hfl' equivalent to 'hl'", {
   expect_equal(my_coef, values_newSB_hfl)
 })
 
+# v930----
+suppressWarnings(
+  createStudy(path = tempdir(), 
+              study_name = "scenariobuilder9.3", 
+              antares_version = "9.3"))
+
+# default area with st cluster
+area_test = "al" 
+lapply(area_test, createArea)
+
+test_that("Add new 'sts' equivalent to 'sct apports'", {
+  createClusterST(area = area_test, 
+                  cluster_name = "default_prop")
+  opts <- simOptions()
+  ldata <- scenarioBuilder(
+    n_scenario = 5,
+    n_mc = 5,
+    areas = area_test
+  )
+  
+  updateScenarioBuilder(ldata = ldata,
+                        series = "sts")
+  
+  newSTS <- readScenarioBuilder(as_matrix = FALSE)
+  expect_true("sts" %in% names(newSTS))
+
+})
+
+test_that("Add new 'hfl' to scenariobuilder of sts", {
+
+  # hfl
+  # nb coeff equivalent to nb areas
+  my_coef <- runif(length(getAreas()))
+  
+  opts <- simOptions()
+  
+  # one constraint => nb areas must be equal to nb coeff
+  ldata <- scenarioBuilder(
+    n_scenario = 10,
+    n_mc = 10,
+    areas = area_test,
+    coef_hydro_levels = my_coef
+  )
+  
+  updateScenarioBuilder(ldata = ldata,
+                        series = "hfl")
+  
+  newSB <- readScenarioBuilder(as_matrix = FALSE)
+  expect_true("hfl" %in% names(newSB))
+  expect_true("sts" %in% names(newSB))
+  
+})
+
+
+test_that("Add new 'sta'equivalent to 'sct contraintes'", {
+  
+  name_no_prefix <- "add_constraints"
+  
+  constraints_properties <- list(
+    "withdrawal-1"=list(
+      variable = "withdrawal",
+      operator = "equal",
+      hours = c("[1,3,5]", 
+                "[120,121,122,123,124,125,126,127,128]")
+    ),
+    "netting-1"=list(
+      variable = "netting",
+      operator = "less",
+      hours = c("[1, 168]")
+    ))
+  
+  # creat cluster
+  createClusterST(area = area_test, 
+                  cluster_name = name_no_prefix, 
+                  constraints_properties = constraints_properties)
+
+  ldata <- scenarioBuilder(
+    n_scenario = 5,
+    n_mc = 5,
+    areas = area_test
+  )
+  
+  updateScenarioBuilder(ldata = ldata,
+                        series = "sta")
+  
+  newSB <- readScenarioBuilder(as_matrix = FALSE)
+  expect_true("sta" %in% names(newSB))
+  
+})
+
+test_that("Add another'sta'to an existing one", {
+  
+  name_no_prefix <- "second_constraints"
+  
+  constraints_properties <- list(
+    "test"=list(
+      variable = "withdrawal",
+      operator = "equal",
+      hours = c("[1,50,20]")
+    ))
+  
+  # creat cluster
+  createClusterST(area = area_test, 
+                  cluster_name = name_no_prefix, 
+                  constraints_properties = constraints_properties)
+  
+  ldata <- scenarioBuilder(
+    n_scenario = 3,
+    n_mc = 3,
+    areas = area_test
+  )
+  
+  updateScenarioBuilder(ldata = ldata,
+                        series = "sta")
+  
+  newSB <- readScenarioBuilder(as_matrix = FALSE)
+  expect_true("sta" %in% names(newSB))
+})
+
 deleteStudy()