+++ NEWS.md
@@ -1,6 +1,26 @@
 > Copyright © 2016 RTE Reseau de transport d’electricite
 
-# antaresEditObject 0.9.2
+
+# antaresEditObject 0.9.3
+(cf. Antares v9.3 changelog)
+
+NEW FEATURES :  
+
+* `createClusterST()`/`editClusterST()` :  
+  - **New properties** (*allow-overflow*, see list of properties according to study version of Antares with `storage_values_default()`) 
+* `createClusterST()`/`editClusterST()` :  
+  - **New dimension of time series** (all matrices will be of (8760, N), noting that N >= 1)
+  - **Suppression of properties** (refreshtimeseries,refreshintervalload, refreshintervalhydro, refreshintervalwind, refreshintervalthermal, refreshintervalsolar from settings/generaldata.ini)
+* `createCluster()` :   
+  - **Dynamic groups** (The group parameter is now dynamic and has no restrictions. The default group value for Renewable depends on the version: Other RES 1 if < 930, Other if >= 930).
+* `updateScenarioBuilder()` New types of series "sts" and "sta" ("sct apports level" and "sct contraintes") is available
+
+BUGFIXES :    
+
+* `createBindingConstraint()/editBindingConstraint()` do not allow to create/edit a binding constraint if a cluster referenced in a coefficient does not exist.
+
+
+# antaresEditObject 0.9.2.9000
 (cf. Antares v9.2 changelog)
 
 NEW FEATURES :